9: DELETE /api/comments/:comment_id endpoint

+++ app/models/articles.models.js
@@ -10,7 +10,7 @@ exports.selectArticleById = (articleId) => {
 exports.selectAllArticles = () => {
 	return db
 		.query(
-			`SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.article_id) AS comment_count
+			`SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.article_id)::INT AS comment_count
   FROM articles 
   LEFT JOIN comments 
   ON articles.article_id = comments.article_id
